Move new methods to v2
